+++ src/aui-scheduler/js/aui-scheduler-view-table.js
@@ -577,6 +577,23 @@ var SchedulerTableView = A.Component.create({
         },
 
         /**
+         * Returns the date interval in which this view shows events for.
+         *
+         * @method getDateInterval
+         * @return {Object} Object with 2 keys: startDate and endDate. Undefined
+         *   keys are interpreted as unlimited sides of the interval.
+         */
+        getDateInterval: function() {
+            var daysAmount = this.getWeekDaysCount() * this._getDisplayRowsCount() - 1,
+                startDate = this._findCurrentIntervalStart();
+
+            return {
+                endDate: DateMath.toLastHour(DateMath.add(startDate, DateMath.DAY, daysAmount)),
+                startDate: DateMath.toMidnight(startDate)
+            };
+        },
+
+        /**
          * Returns the list of all events that intersect with a given date.
          *
          * @method getIntersectEvents
@@ -636,6 +653,18 @@ var SchedulerTableView = A.Component.create({
         },
 
         /**
+         * Get the number of days that should be shown in the week.
+         *
+         * @method getWeekDaysCount
+         * @return {Number}
+         */
+        getWeekDaysCount: function() {
+            var displayDaysInterval = this.get('displayDaysInterval');
+
+            return Math.min(displayDaysInterval, WEEK_LENGTH);
+        },
+
+        /**
          * Hides this `SchedulerViewTable` event's `overlay` component.
          *
          * @method hideEventsOverlay
@@ -688,8 +717,7 @@ var SchedulerTableView = A.Component.create({
 
             instance.bodyNode.all('.' + CSS_SVT_TABLE_DATA).remove();
 
-            var displayDaysInterval = instance.get('displayDaysInterval');
-            var weekDaysCount = Math.min(displayDaysInterval, WEEK_LENGTH);
+            var weekDaysCount = this.getWeekDaysCount();
 
             var finalDate = DateMath.add(startDateRef, DateMath.DAY, (weekDaysCount * this.tableRows.size()) - 1);
             instance._findIntersections(startDateRef, finalDate);